Revert "Update to java 25 (#478)"
This reverts commit 82202c3557ca952c80b66b1e1e9eddd223bc717d.

+++ generator/src/main/java/com/faforever/neroxis/generator/terrain/PathedPlateauLastTerrainGenerator.java
@@ -6,7 +6,7 @@ public abstract class PathedPlateauLastTerrainGenerator extends BasicLastTerrain
     @Override
     protected void plateausSetup() {
         int mapSize = map.getSize();
-        int maxStepSize = mapSize / 128;
+        float maxStepSize = mapSize / 128f;
         int maxMiddlePoints = 16;
         int numPaths = (int) (12 * plateauDensity) / symmetrySettings.spawnSymmetry().getNumSymPoints();
         int bound = 0;
@@ -14,7 +14,7 @@ public abstract class PathedPlateauLastTerrainGenerator extends BasicLastTerrain
 
         MapMaskMethods.pathInCenterBounds(random.nextLong(), plateaus, maxStepSize, numPaths, maxMiddlePoints, bound,
                                           (float) (StrictMath.PI / 2));
-        plateaus.inflate(mapSize / 256).setSize(mapSize / 4);
+        plateaus.inflate(mapSize / 256f).setSize(mapSize / 4);
         plateaus.dilute(.5f, 4).setSize(mapSize + 1);
         plateaus.blur(12);
     }
